[refactor] 테스트(Unit Test 아님) 관련 model core:data:test:test-api 모듈로 이전
