+++ core/data/onboard/onboard-api/src/main/java/com/qriz/app/core/data/onboard/onboard_api/repository/OnBoardRepository.kt
@@ -1,12 +1,13 @@
 package com.qriz.app.core.data.onboard.onboard_api.repository
 
+import com.qriz.app.core.data.test.test_api.model.Option
 import com.qriz.app.core.data.onboard.onboard_api.model.PreCheckConcept
-import com.qriz.app.core.data.onboard.onboard_api.model.Test
+import com.qriz.app.core.data.test.test_api.model.Test
 
 interface OnBoardRepository {
     fun submitSurvey(concepts: Collection<PreCheckConcept>)
 
     suspend fun getPreviewTest(): Test
 
-    fun submitPreview(answer: Map<Long, String>)
+    fun submitPreview(answer: Map<Long, Option>)
 }
